Move Entrypoint and SkipExternalService from core to helpers
Their only uses are in the helpers package, so move them there.
