+++ skipruntime-ts/core/src/index.ts
@@ -51,38 +51,10 @@ import {
 } from "./binding.js";
 
 export { UnknownCollectionError, sk_freeze, isSkManaged };
-export { SkipExternalService } from "./remote.js";
 export { Sum, Min, Max, Count } from "./utils.js";
 
 export type JSONMapper = Mapper<Json, Json, Json, Json>;
 export type JSONLazyCompute = LazyCompute<Json, Json>;
-
-/**
- * An entry point of a Skip reactive service.
- *
- * URLs for the service's control and streaming APIs can be constructed from an `Entrypoint`.
- */
-export type Entrypoint = {
-  /**
-   * Hostname of the service.
-   */
-  host: string;
-
-  /**
-   * Port to use for the service's streaming interface.
-   */
-  streaming_port: number;
-
-  /**
-   * Port to use for the service's control interface.
-   */
-  control_port: number;
-
-  /**
-   * Flag that when set indicates that https should be used instead of http.
-   */
-  secured?: boolean;
-};
 
 class Handles {
   private nextID: number = 1;
