+++ skipruntime-ts/helpers/src/index.ts
@@ -9,5 +9,6 @@ export {
   GenericExternalService,
   Polled,
 } from "./external.js";
+export { SkipExternalService } from "./remote.js";
 export { SkipServiceBroker, fetchJSON, type Entrypoint } from "./rest.js";
-export { Count, Max, Min, SkipExternalService, Sum } from "@skipruntime/core";
+export { Count, Max, Min, Sum } from "@skipruntime/core";
